refact: split main in 2 files; chore: add engine version
