+++ .gitignore
@@ -43,5 +43,6 @@ out
 
 # Generated files
 imgui.ini
+app.ini
 recorded_events.json
 logs
