+++ src/Engine/include/Engine/details/Version.hpp.in
@@ -1,0 +1,10 @@
+#pragma once
+
+#define PROJECT_NAME    "@PROJECT_NAME@"    /*!< @brief Name of the project */
+
+#define PROJECT_VERSION "@PROJECT_VERSION@" /*!< @brief Full Version as string */
+
+#define PROJECT_VERSION_MAJOR @PROJECT_VERSION_MAJOR@ /*!< @brief Version Major */
+#define PROJECT_VERSION_MINOR @PROJECT_VERSION_MINOR@ /*!< @brief Version Minor */
+#define PROJECT_VERSION_PATCH @PROJECT_VERSION_PATCH@ /*!< @brief Version Patch */
+//#define PROJECT_VERSION_TWEAK @PROJECT_VERSION_TWEAK@ /*!< @brief Version Tweak */
